feat: add new API for perform check on reconciled events (#3)
This PR contains a new feature made in order to provide a new tool during error tracing in production environment. This new API permits to check if a passed event (referenced by `PartitionKey`, `RowKey` and event `timestamp`) was correctly reconciled, i.e. is present either in hot and cold storages.  
This API was defined in order to be used with alerting message: the log message containing these three information can be used for this checks. 

#### List of Changes
 - Add new API for execute reconciliation checks on events